requirements/base.txt
python-memcached
django-sekizai
django-statici18n
django_compressor
csscompressor
beautifulsoup4
lxml
django-htmlmin
-e git+git@github.com:nargeel/django-fancy-imagefield.git@master#egg=django-fancy-imagefield
-e git+git@github.com:nargeel/django-contrib.git@master#egg=django-contrib